Corrige compositor, paginas y notificaciones del P4
Serializa cada transferencia DMA2D con su callback para evitar cuadros mezclados y elimina el presenter que podia bloquear loopTask hasta TASK_WDT. Reduce el cache de pagina a la banda movil, limita animaciones pesadas a 30 fps y cede CPU durante composiciones largas.

Limpia la isla antes de abrir la caja sin perder la cola, refresca el Home al cambiar la personalizacion y corrige el borrado parcial del resultado local de Cowork. Anade regresiones de compositor, pagina y notificacion.

+++ FlexOS_AI.cpp
@@ -480,8 +480,8 @@ static uint8_t aiDoQuery(const char* reqBody, size_t reqLen, AiResult* res){
       }
       if(rc != AI_ERR_TIMEOUT){
         resp[got] = 0;
-        rc = aiParseResponse(resp, got, res) ? AI_ERR_NONE
-                                             : (res->err ? res->err : AI_ERR_PARSE);
+        rc = aiParseResponse(resp, got, res) ? (uint8_t)AI_ERR_NONE
+                                             : (res->err ? res->err : (uint8_t)AI_ERR_PARSE);
       }
     }
     http.end();
